+++ src/stories/CardContainer.stories.ts
@@ -19,6 +19,7 @@ type Story = StoryObj<typeof meta>;
 // More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Primary: Story = {
   args: {
+    isLeader: true,
     stack: ['React.js', 'Angular', 'Django', 'GraphQL', 'Laravel'],
     languages: ['js', 'python', 'jdas'],
     data: {
