+++ src/widgets/Card/types.ts
