+++ src/widgets/Card/ui/Name.tsx
@@ -2,7 +2,7 @@ import { ReactElement, ReactNode } from 'react';
 import styles from './Card.module.scss';
 import { flags } from '@/shared/const/flags';
 import Image from 'next/image';
-import { NameProps } from '../Card';
+import { NameProps } from '../types';
 
 export function Name({ name, country }: NameProps) {
   const Icon = flags[country];
