+++ src/widgets/Card/ui/Position.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 import styles from './Card.module.scss';
-import { PositionProps } from '../Card';
+import { PositionProps } from '../types';
 
 export function Position({ children }: PositionProps) {
   return <span className={styles.position}> {children} </span>;
